+++ app/components/DBLEntryRow.js
@@ -368,13 +368,18 @@ class DBLEntryRow extends PureComponent<Props> {
     return (null);
   };
 
-  pickBackgroundColor = () => {
+  pickBackgroundColor = (isForRow) => {
     const {
       classes, status, revision, parent, dblId
     } = this.props;
-    const effectiveRevision = bundleService.getRevisionOrParentRevision(dblId, revision, parent);
     switch (status) {
-      case 'DRAFT': return effectiveRevision ? classes.draftRevision : classes.draftNew;
+      case 'DRAFT': {
+        if (isForRow) {
+          return classes.storedMode;
+        }
+        const effectiveRevision = bundleService.getRevisionOrParentRevision(dblId, revision, parent);
+        return effectiveRevision ? classes.draftRevision : classes.draftNew;
+      }
       case 'NOT_STARTED': return classes.noneStoredMode;
       default:
         return classes.storedMode;
@@ -406,10 +411,11 @@ class DBLEntryRow extends PureComponent<Props> {
     }
     const resourceManagerMode = status === 'DRAFT' ? 'addFiles' : 'download';
     const laterEntryRevisionsCount = this.props.laterEntryRevisions.length;
-    const laterRevisionsBadge = laterEntryRevisionsCount ? `${laterEntryRevisionsCount}` : '';
+    const laterRevisionsBadge = laterEntryRevisionsCount ? `${laterEntryRevisionsCount}+` : '';
+    const mediumIconMarginRight = ux.getMediumIcon(medium);
     return (
       <div
-        className={classNames(styles.bundleRow, this.pickBackgroundColor())}
+        className={classNames(styles.bundleRow, this.pickBackgroundColor(true))}
         key={bundleId}
         onKeyPress={this.onKeyPress}
         onClick={this.onClickBundleRow}
@@ -420,7 +426,7 @@ class DBLEntryRow extends PureComponent<Props> {
         <div className={styles.bundleRowTop}>
           <div className={styles.bundleRowTopLeftSideIcon}>
             <Tooltip title={medium}>
-              { ux.getMediumIcon(medium) }
+              { mediumIconMarginRight }
             </Tooltip>
           </div>
           <div className={styles.bundleRowTopLeftSideLanguageAndCountry}>
@@ -432,9 +438,9 @@ class DBLEntryRow extends PureComponent<Props> {
           <div className={styles.bundleRowTopMiddle}>
             <Tooltip title="Switch revision">
               <Button
-                variant="text"
+                variant="outlined"
                 size="small"
-                className={classes.button}
+                className={classNames(classes.button, this.pickBackgroundColor())}
                 disabled={dblId === undefined}
                 onClick={this.onClickManageResources('revisions')}
               >
@@ -463,11 +469,16 @@ class DBLEntryRow extends PureComponent<Props> {
           </div>
           <div className={styles.bundleRowTopRightSide}>
             {this.showStoredButton() && (
-              <Button variant="text" size="small" className={classes.button}
+              <Button
+                variant="text"
+                size="small"
+                className={classNames(classes.button, this.pickBackgroundColor())}
                 onClick={this.onClickManageResources(resourceManagerMode)}
               >
                 <ControlledHighlighter {...this.getHighlighterSharedProps(displayAs.status)} />
-                <Folder className={classNames(classes.rightIcon, classes.iconSmall)} />
+                <Badge badgeContent={ux.getMediumIcon(medium, { className: classNames(classes.rightIcon, classes.iconSmaller) })} >
+                  <Folder className={classNames(classes.rightIcon, classes.iconSmall)} />
+                </Badge>
               </Button>
             )}
             {this.showStatusAsText() && (
@@ -476,7 +487,10 @@ class DBLEntryRow extends PureComponent<Props> {
               </div>
             )}
             {this.showDownloadButton() && (
-              <Button variant="text" size="small" className={classes.button}
+              <Button
+                variant="outlined"
+                size="small"
+                className={classes.button}
                 onKeyPress={this.onClickManageResources('download')}
                 onClick={this.onClickManageResources('download')}
               >
@@ -495,7 +509,7 @@ class DBLEntryRow extends PureComponent<Props> {
           </div>
         )}
         {isSelected && (
-          <Toolbar style={{ minHeight: '36px' }}>
+          <Toolbar style={{ minHeight: '36px' }} className={this.pickBackgroundColor()}>
             {this.shouldShowEdit() &&
             <Button
               disabled={this.shouldDisableDraftRevisionOrEdit()}
@@ -552,7 +566,7 @@ class DBLEntryRow extends PureComponent<Props> {
                     key={mediumOption}
                     onClick={this.handleClickMediaType(mediumOption)}
                   >
-                    {ux.getMediumIcon(mediumOption)}
+                    {mediumIconMarginRight}
                     {mediumOption}
                   </MenuItem>
                 ))}
@@ -609,6 +623,9 @@ const materialStyles = theme => ({
   },
   iconSmall: {
     fontSize: 20,
+  },
+  iconSmaller: {
+    fontSize: 10,
   },
   draftRevision: { backgroundColor: lighten(theme.palette.secondary.light, 0.85) },
   draftNew: { backgroundColor: lighten(theme.palette.primary.main, 0.60) },