+++ app/components/EnhancedTable.js
@@ -33,14 +33,10 @@ type Props = {
   columnConfig: [],
   defaultOrderBy: string,
   secondarySorts: [],
-  selectAll: boolean,
+  selectedIds: [],
   multiSelections?: boolean,
   customSorts?: {},
-  onSelectedRowIds: () => {},
-  handleAddByFile?: () => {},
-  handleAddByFolder?: () => {},
-  getSuggestions?: () => {},
-  onAutosuggestInputChanged?: () => {}
+  onSelectedRowIds: () => {}
 };
 
 function getDataRowIds(data) {
@@ -51,26 +47,18 @@ function getSelectableData(data) {
   return data.filter(d => !d.disabled);
 }
 
-function getAllSelectableRowIds(data) {
-  const selectableData = getSelectableData(data);
-  return getDataRowIds(selectableData);
-}
-
 class EnhancedTable extends Component<Props> {
   props: Props;
   state = {
     order: this.props.orderDirection,
     orderBy: this.props.defaultOrderBy || this.props.columnConfig[0].name,
-    selectedRowIds: []
+    selectedRowIds: this.props.selectedIds
   };
 
   componentWillReceiveProps(nextProps) {
     // You don't have to do this check first, but it can help prevent an unneeded render
-    if (nextProps.multiSelections && nextProps.selectAll && nextProps.data.length &&
-      (this.props.data !== nextProps.data ||
-      this.state.selectedRowIds.length === 0 ||
-      nextProps.selectAll !== this.props.selectAll)) {
-      this.setState({ selectedRowIds: getAllSelectableRowIds(nextProps.data) });
+    if (nextProps.selectedIds !== this.props.selectedIds) {
+      this.setState({ selectedRowIds: nextProps.selectedIds });
     }
   }
 
@@ -210,17 +198,10 @@ class EnhancedTable extends Component<Props> {
 
   render() {
     const { classes } = this.props;
-    const { selectedRowIds, orderBy, order } = this.state;
+    const { orderBy, order } = this.state;
 
     return (
       <Paper className={classes.root}>
-        <EnhancedTableToolbar
-          numSelected={selectedRowIds.length}
-          handleAddByFile={this.props.handleAddByFile}
-          handleAddByFolder={this.props.handleAddByFolder}
-          getSuggestions={this.props.getSuggestions}
-          onAutosuggestInputChanged={this.props.onAutosuggestInputChanged}
-        />
         <MuiTable
           data={this.getSortedData()}
           columns={this.columns()}
@@ -246,11 +227,7 @@ class EnhancedTable extends Component<Props> {
 EnhancedTable.defaultProps = {
   orderDirection: 'asc',
   customSorts: {},
-  multiSelections: false,
-  handleAddByFile: undefined,
-  handleAddByFolder: undefined,
-  getSuggestions: undefined,
-  onAutosuggestInputChanged: undefined
+  multiSelections: false
 };
 
 export default withStyles(styles)(EnhancedTable);