+++ app/components/EnhancedTableToolbar.js
@@ -49,10 +49,11 @@ const toolbarStyles = theme => ({
 type Props = {
   classes: {},
   numSelected: number,
-  handleAddByFile: ?() => {},
-  handleAddByFolder: ?() => {},
-  getSuggestions: ?() => {},
-  onAutosuggestInputChanged: ?() => {}
+  enableEditContainer?: boolean,
+  handleAddByFile?: () => {},
+  handleAddByFolder?: () => {},
+  getSuggestions?: () => {},
+  onAutosuggestInputChanged?: () => {}
 };
 
 class EnhancedTableToolbar extends Component<Props> {
@@ -80,7 +81,7 @@ class EnhancedTableToolbar extends Component<Props> {
   }
 
   render() {
-    const { numSelected, classes, handleAddByFile } = this.props;
+    const { numSelected, classes, handleAddByFile, enableEditContainer } = this.props;
     const { anchorEl } = this.state;
     return (
       <Toolbar
@@ -99,7 +100,7 @@ class EnhancedTableToolbar extends Component<Props> {
         </div>
         <div className={classes.spacer} />
         <div style={{ width: 700 }}>
-          {handleAddByFile && numSelected > 0 ? (
+          {enableEditContainer ? (
             <IntegrationAutosuggest
               getSuggestions={this.props.getSuggestions}
               onInputChanged={this.props.onAutosuggestInputChanged}
@@ -108,7 +109,7 @@ class EnhancedTableToolbar extends Component<Props> {
         <div className={classes.actions}>
           {handleAddByFile ? (
             <div>
-              <Tooltip title="Add resource">
+              <Tooltip title="Add resource(s)">
                 <Button
                   aria-owns={anchorEl ? 'simple-menu' : null}
                   aria-haspopup="true"
@@ -153,5 +154,13 @@ class EnhancedTableToolbar extends Component<Props> {
     );
   }
 }
+
+EnhancedTableToolbar.defaultProps = {
+  enableEditContainer: false,
+  handleAddByFile: undefined,
+  handleAddByFolder: undefined,
+  getSuggestions: undefined,
+  onAutosuggestInputChanged: undefined
+};
 
 export default withStyles(toolbarStyles)(EnhancedTableToolbar);